Spec 24334: Add workflow views for document abstraction - add permissions handler and register roles and permissions
