+++ src/org/labkey/workflow/PermissionsHandlerImpl.java
@@ -24,75 +24,10 @@ import java.util.Set;
  */
 public class PermissionsHandlerImpl extends PermissionsHandler
 {
-    protected final boolean _hasAdmin;
 
     public PermissionsHandlerImpl(@NotNull User user, @NotNull Container container)
     {
         super (user, container);
-        _hasAdmin = _container.hasPermission(_user, AdminPermission.class);
-    }
-
-    @Override
-    public boolean canStartProcess(@NotNull String processDefinitionKey)
-    {
-        return _container.hasPermission(_user, ReadPermission.class);
-    }
-
-    @Override
-    public boolean canDeployProcess(@NotNull String processDefinitionKey) { return _hasAdmin; }
-
-    @Override
-    public boolean canView(@NotNull WorkflowProcess process)
-    {
-        return (process.getInitiatorId() == _user.getUserId()) || _hasAdmin;
-    }
-
-    @Override
-    public boolean canAccessData(@NotNull WorkflowProcess process)
-    {
-        return _hasAdmin;
-    }
-
-    @Override
-    public boolean canDelete(@NotNull WorkflowProcess process)
-    {
-        return process.getInitiatorId() == _user.getUserId() || _hasAdmin;
-    }
-
-    @Override
-    public boolean canClaim(@NotNull WorkflowTask task)
-    {
-        return _hasAdmin || task.isInCandidateGroups(_user);
-    }
-
-    @Override
-    public boolean canDelegate(@NotNull WorkflowTask task)
-    {
-        return _hasAdmin;
-    }
-
-    @Override
-    public boolean canAssign(@NotNull WorkflowTask task)
-    {
-        return _hasAdmin;
-    }
-
-    @Override
-    public boolean canView(@NotNull WorkflowTask task)
-    {
-        return ((task.getAssignee() != null && task.getAssignee().getUserId() == _user.getUserId())) || canClaim(task) || canDelegate(task);
-    }
-
-    @Override
-    public boolean canAccessData(@Nullable WorkflowTask task)
-    {
-        return _hasAdmin;
-    }
-
-    @Override
-    public boolean canComplete(@NotNull WorkflowTask task)
-    {
-        return task.isAssigned(_user);
     }
 
     @Override
